+++ pages/checkout-form.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <link rel="stylesheet" href="./style/style.css">
+    <link rel="stylesheet" href="../style/style.css">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
@@ -93,6 +93,6 @@
         </div>
     </div>
 </div>
-<script type="module" src="script/FormController.js"></script>
+<script type="module" src="../script/FormController.js"></script>
 </body>
 </html>
