+++ pages/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <link rel="stylesheet" href="./style/style.css">
+    <link rel="stylesheet" href="../style/style.css">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
@@ -10,11 +10,14 @@
     <title>Book_shop</title>
 </head>
     <body class="body">
-    <div class="wrapper"></div>
-    <script type="module" src="script/Index.js"></script>
-    <script type="module" src="script/ExpandBtnModal.js"></script>
-    <script type="module" src="script/CartController.js"></script>
-    <script type="module" src="script/FormController.js"></script>
+    <div class="project-wrapper">
+        <div class="wrapper"></div>
+    </div>
+
+    <script type="module" src="../script/Index.js"></script>
+    <script type="module" src="../script/ExpandBtnModal.js"></script>
+    <script type="module" src="../script/CartController.js"></script>
+    <script type="module" src="../script/FormController.js"></script>
 
     </body>
 </html>
