+++ script/CreateElement.js
@@ -1,4 +1,4 @@
-export  default function createElement(tag, ...classList) {
+export default function createElement(tag, ...classList) {
     const element = document.createElement(tag);
     element.classList.add(...classList);
     return element;
