+++ style/style.css
@@ -11,8 +11,14 @@
 
 .body{
     margin: 0;
-}
 
+}
+.project-wrapper{
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    overflow-x: hidden;
+}
 .wrapper{
     width: 1270px;
     margin: auto;
